scoreDistributionTable중복정의수정
