+++ src/features/passAnalysis/components/PassAnalysisMenu/PracticalScoreAnalysisMenu.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 import styled from '@emotion/styled';
 import { Tooltip } from '@chakra-ui/react';
 import PracticalScoreAnalysisMenuItem from './PracticalScoreAnalysisMenuItem';
-import DistributionTableContainer from '@/components/Table/ScoreDistributionTableContainer';
+import DistributionTableContainer from '@/container/PracticalScoreDistributionChartContainer';
 import Info from '../../../../../public/images/icons/info.svg';
 import HelpOutline from '../../../../../public/images/icons/helpOutline.svg';
 import ChangePracticalScoreModal from '../modal/ChangePracticalScoreModal';
